refactor: :building_construction: Move the extraction modules into a subfolder 'extractors'
